CharactersBehaviour has characterInfo
CharactersBehaviours generated by the LevelScript now inherit one of the
three characterInfos in the GlobalGameManager script. It extracts the
relevant data (movement, attack range, attack damage, name) and uses
this data now. Changes to the characterInfo in the characterInfos array
in global will change effectively change a character for the battle as
well.