Working Save/Load
NOTE:
Note #1: ALL card classes must be serializable, as shown now.
Note B: We can not store Color in classes labeled serialized now. Unlike
what the unity documentation says, the color class is not serializable.
I've mended the only class where this is an issue (AoE) by storing the
attack color in the map file instead.

The code is easily extendable to many saves, but currently saves only
one gamesave. Later, we can extend this out to a final-fantasy or
elderscrolls style of saving, with many saves, and choosing between
them. But that requires extra menus that I didn't want to fuss with
right now.