Enemy Behaviour Update
Added enemy behaviour for Patrol and Beserk. The Patrol enemy behaviour
will walk towards a square specified, then return to it's starting
square. There is some leeway in the algorithm that I need to tweak, so
it isn't perfect at the moment.
The Beserk enemy will walk towards the enemy specified. I specified this
in the Level file every time the enemy's turn comes up. The current
implementation of enemies has them running towards the nearest enemy if
there is one, or the 'first' character (as spawned in).