Animation beginnings
** READ **
We CAN NOT make custom animation controllers in unity reliably. I spent
2 hours on this, to find it isn't fully supported. This means we need to
make an animator controller for each animation set. In Animations I have
white_walk which demonstrates the proper way to do this, with the code
I've set up. It's important to note the settings, so there is no
blending in each transition. The code will look for <spriteName>_walk to
set as it's controller. If it can't find it, it fails silently.

I did not change or fiddle with the size of the characters on the board.
This very likely needs to change.

Changed the characterCard to have a sprite name, and changed the
customization screen to display the sprite of said name. This will
likely require tweaking, because I specifically grab the 18th frame,
guessing it will be the same for all characters. It might not be.