Fixed attacking and other goodies
Fixed so you can attack again. Whoops. Not being able to attack through
walls now looks better, and functions a tish better.

Added a defense buff prefab that currently spawns on defense buff, as
well as at the end of every turn that a character takes who has a
defense buff. Wasn't sure how to make this work for only on-cast, or if
we could make it work at the beginning of every turn.